+++ hgraph_unit_tests/_wiring/test_map.py
@@ -26,7 +26,6 @@ from hgraph import (
     debug_print,
     switch_,
     nothing,
-    default,
     Removed,
     sum_,
 )
@@ -263,91 +262,6 @@ def _test_tsl_map(map_test):
 
 
 @pytest.mark.parametrize(
-    ["inputs", "expected"],
-    [
-        [[None, {"a": 1}, {"a": REMOVE_IF_EXISTS}], [0, 1, 0]],
-        [[None, {"a": 1}, {"b": 2}, {"b": REMOVE_IF_EXISTS}, {"a": REMOVE_IF_EXISTS}], [0, 1, 3, 1, 0]],
-        [[{"a": 1, "b": 2, "c": 3}, {"b": REMOVE_IF_EXISTS}, {"a": REMOVE_IF_EXISTS}], [6, 4, 3]],
-        [[{"a": 1}, {"b": 2}, {"c": 3}, {"d": 4}, {"e": 5}], [1, 3, 6, 10, 15]],
-        [
-            [
-                {(chr(ord("a") + i)): i for i in range(26)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(26)},
-            ],
-            [325, 0],
-        ],
-        [
-            [
-                {(chr(ord("a") + i)): i for i in range(26)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(20)},
-                {(chr(ord("a") + i)): i for i in range(20)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(26)},
-                {(chr(ord("a") + i)): i for i in range(26)},
-            ],
-            [325, 135, 325, 0, 325],
-        ],
-    ],
-)
-def test_tsd_reduce(inputs, expected):
-    @graph
-    def reduce_test(tsd: TSD[str, TS[int]]) -> TS[int]:
-        return reduce(add_, tsd, 0)
-
-    assert eval_node(reduce_test, inputs) == expected
-
-
-@pytest.mark.parametrize(
-    ["inputs", "expected"],
-    [
-        [[None, {"a": 1}, {"a": REMOVE_IF_EXISTS}], [0, 1, 0]],
-        [[None, {"a": 1}, {"b": 2}, {"b": REMOVE_IF_EXISTS}, {"a": REMOVE_IF_EXISTS}], [0, 1, 3, 1, 0]],
-        [[{"a": 1, "b": 2, "c": 3}, {"b": REMOVE_IF_EXISTS}, {"a": REMOVE_IF_EXISTS}], [6, 4, 3]],
-        [[{"a": 1}, {"b": 2}, {"c": 3}, {"d": 4}, {"e": 5}], [1, 3, 6, 10, 15]],
-        [
-            [
-                {(chr(ord("a") + i)): i for i in range(26)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(26)},
-            ],
-            [325, 0],
-        ],
-        [
-            [
-                {(chr(ord("a") + i)): i for i in range(26)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(20)},
-                {(chr(ord("a") + i)): i for i in range(20)},
-                {(chr(ord("a") + i)): REMOVE_IF_EXISTS for i in range(26)},
-                {(chr(ord("a") + i)): i for i in range(26)},
-            ],
-            [325, 135, 325, 0, 325],
-        ],
-    ],
-)
-def test_tsd_reduce_no_zero(inputs, expected):
-    @graph
-    def reduce_test(tsd: TSD[str, TS[int]]) -> TS[int]:
-        return reduce(add_, tsd)
-
-    assert eval_node(reduce_test, inputs) == expected
-
-
-@pytest.mark.parametrize(
-    ["inputs", "size", "expected"],
-    [
-        [[None, {0: 1}, None, {1: 2}], Size[2], [0, 1, None, 3]],
-        [[None, {0: 1, 3: 4}, {1: 2, 2: 3}], Size[4], [0, 5, 10]],
-        [[None, {0: 1, 3: 4}, {1: 2, 2: 3}, {4: 8}], Size[5], [0, 5, 10, 18]],
-        [[None, {0: 1, 3: 4}, {1: 2, 2: 3}, {4: 8, 5: 9}], Size[6], [0, 5, 10, 27]],
-    ],
-)
-def test_tsl_reduce(inputs, size, expected):
-    @graph
-    def reduce_test(tsl: TSL[TS[int], SIZE]) -> TS[int]:
-        return reduce(add_, tsl, 0)
-
-    assert eval_node(reduce_test, inputs, resolution_dict={"tsl": TSL[TS[int], size]}) == expected
-
-
-@pytest.mark.parametrize(
     ["inputs", "size", "expected"],
     [
         [[None, {0: 1}, None, {1: 2}], Size[2], [0, 1, None, 3]],
@@ -514,11 +428,3 @@ def test_map_preexisting_keys():
     ]
 
 
-def test_reduce_map():
-    @graph
-    def g(items: TSD[int, TSD[int, TS[int]]]) -> TSD[int, TS[int]]:
-        return items.reduce(lambda x, y: map_(lambda i, j: default(i, 0) + default(j, 0), x, y))
-
-    assert eval_node(
-        g, [{1: {1: 1, 2: 2}}, {2: {1: 3, 2: 4}}, {3: {2: 1, 3: 3}}], __trace__={"start": False, "stop": False}
-    ) == [{1: 1, 2: 2}, {1: 4, 2: 6}, {1: 4, 2: 7, 3: 3}]